% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrywise_simulation.R
\name{binary_simulation}
\alias{binary_simulation}
\title{Matrix generation from Bernoulli random variables}
\usage{
binary_simulation(prob_mat)
}
\arguments{
\item{prob_mat}{probability matrix}
}
\value{
a binary matrix
}
\description{
Matrix generation from Bernoulli random variables
}
