% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{bray_curtis_distance}
\alias{bray_curtis_distance}
\title{Calculate the Bray-Curtis distance between true composition and estimated composition}
\usage{
bray_curtis_distance(truth, estimation)
}
\arguments{
\item{truth}{true abundance matrix}

\item{estimation}{estimated abundance matrix}
}
\value{
a scalar representing average bray-curtis distance
}
\description{
Calculate the Bray-Curtis distance between true composition and estimated composition
}
