% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composition_utils.R
\name{clr_transform}
\alias{clr_transform}
\title{centered log ratio transformation}
\usage{
clr_transform(count_mat, impute_scale = 0.5)
}
\arguments{
\item{count_mat}{count matrix (nsample*nfeature)}

\item{impute_scale}{for an entry with zeros, impute the zeros with scale* smallest nonzero value in the column
@return clr transformed matrix}
}
\description{
centered log ratio transformation
}
