% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_coda_lasso.R
\name{codalasso_lambda_tune}
\alias{codalasso_lambda_tune}
\title{Select best lambda for codalasso based on cross validation}
\usage{
codalasso_lambda_tune(
  y,
  X,
  lambdas = seq(0.05, 0.55, 0.1),
  nfolds = 5,
  seed = 1
)
}
\arguments{
\item{y}{response variable}

\item{X}{predictor variable matrix}

\item{lambdas}{penalty parameter range}

\item{nfolds}{number of folds for cross validation}

\item{seed}{random seed for train validation split}
}
\value{
optimal lambda
}
\description{
Select best lambda for codalasso based on cross validation
}
