% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{coef_var}
\alias{coef_var}
\title{Calculate coefficient of variation of relative abundance for all the features}
\usage{
coef_var(truth, estimation)
}
\arguments{
\item{truth}{true abundance matrix}

\item{estimation}{estimated abundance matrix}
}
\value{
a scalar representing average coefficient of variation difference
}
\description{
Calculate coefficient of variation of relative abundance for all the features
}
