% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_coda_lasso.R
\name{cv.codalasso}
\alias{cv.codalasso}
\title{split into train and validation multiple times}
\usage{
cv.codalasso(y, X, lambda, nfolds = 5, seed = 1)
}
\arguments{
\item{y}{response variable}

\item{X}{predictor variable matrix}

\item{lambda}{penalty parameter}

\item{nfolds}{number of folds for cross validation}

\item{seed}{random seed for train validation split}
}
\value{
list of training AUC, validation AUC and coefficients of all runs
}
\description{
split into train and validation multiple times
}
