% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composition_utils.R
\name{dist_composition}
\alias{dist_composition}
\title{Calculate distance between sample compositions and carry out statistical tests}
\usage{
dist_composition(count_mat, metric = "bray", nperm = 999, covariate = NULL)
}
\arguments{
\item{count_mat}{count matrix}

\item{metric}{metric for calculating sample distance, by default bray-curtis}

\item{nperm}{number of permutations needed for permanova test}

\item{covariate}{vector of a covariate for permanova test}
}
\value{
a list of distance matrix, MDS coordinates and permanova test result
}
\description{
Calculate distance between sample compositions and carry out statistical tests
}
