% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{feature_correlation}
\alias{feature_correlation}
\title{Calculate average correlation of relative abundance for all the features}
\usage{
feature_correlation(truth, estimation)
}
\arguments{
\item{truth}{true abundance matrix}

\item{estimation}{estimated abundance matrix}
}
\value{
a scalar representing average correlation with truth
}
\description{
Calculate average correlation of relative abundance for all the features
}
