% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_coda_lasso.R
\name{fit_codalasso}
\alias{fit_codalasso}
\title{Fit coda lasso multiple times with random train/test splitting. Cross validation involved}
\usage{
fit_codalasso(
  y_train,
  X_train,
  y_test,
  X_test,
  lambdas = seq(0.05, 0.55, 0.1),
  seed = 1
)
}
\arguments{
\item{y_train}{response variable of training data}

\item{X_train}{predictor variable matrix of training data}

\item{y_test}{response variable of test data}

\item{X_test}{predictor variable matrix of test data}

\item{lambdas}{range of lambdas to try}

\item{seed}{random seed that decides train test split}
}
\value{
list of training AUC, test AUC and coefficients of all runs
}
\description{
Fit coda lasso multiple times with random train/test splitting. Cross validation involved
}
