% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_lasso.R
\name{fit_logistic_lasso}
\alias{fit_logistic_lasso}
\title{Fit LASSO logistic regression}
\usage{
fit_logistic_lasso(input, output, train_prop = 0.7, seed = 1)
}
\arguments{
\item{input}{input predictor matrix}

\item{output}{binary response}

\item{train_prop}{proportion of data as training}

\item{seed}{random seed for spliting data into train and test}
}
\value{
a list with coefficients and train AUC, test AUC
}
\description{
Fit LASSO logistic regression
}
