% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RF.R
\name{fit_rf}
\alias{fit_rf}
\title{Randomly split the data into train/test and apply random forest model for prediction}
\usage{
fit_rf(X, y, seed = 20)
}
\arguments{
\item{X}{predictor matrix}

\item{y}{response variable}

\item{seed}{random seed for spliting train/test data}
}
\value{
a list with the training AUROC, test AUROC and the variable importance
}
\description{
Randomly split the data into train/test and apply random forest model for prediction
}
