% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{frobenius_norm}
\alias{frobenius_norm}
\title{Calculate the frobenious norm error between two composition matrix}
\usage{
frobenius_norm(truth, estimation)
}
\arguments{
\item{truth}{true abundance matrix}

\item{estimation}{estimated abundance matrix}
}
\value{
a scalar representing frobenious norm error
}
\description{
Calculate the frobenious norm error between two composition matrix
}
