% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrywise_simulation.R
\name{gamma_simulation}
\alias{gamma_simulation}
\title{Matrix generation from Gamma distribution}
\usage{
gamma_simulation(mean_mat, scale = 1)
}
\arguments{
\item{mean_mat}{mean of gamma distribution for each entry}

\item{scale}{scale of gamma distribution for all entries}
}
\value{
matrix of values from gamma dstributions
}
\description{
Matrix generation from Gamma distribution
}
