% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{KL_divergence}
\alias{KL_divergence}
\title{Calculate the KL divergence between true composition and estimated composition}
\usage{
KL_divergence(truth, estimation)
}
\arguments{
\item{truth}{true abundance matrix}

\item{estimation}{estimated abundance matrix}
}
\value{
a scalar representing average KL divergence of sample compositions
}
\description{
Calculate the KL divergence between true composition and estimated composition
}
