% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composition_utils.R
\name{normalize}
\alias{normalize}
\title{Normalize counts for each row of samples}
\usage{
normalize(count_mat, libsize = 1)
}
\arguments{
\item{count_mat}{count matrix (nsample*nfeature)}

\item{libsize}{library size to normalize at}
}
\value{
normalized count matrix
}
\description{
Normalize counts for each row of samples
}
