% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFBM_utils.R
\name{plot_binary_mat}
\alias{plot_binary_mat}
\title{Visualize binary masks and save as png files}
\usage{
plot_binary_mat(
  count_mat,
  thresholds,
  folder,
  cmap = colorRamp2(c(0, 1), c("white", "#2c7cc7")),
  include_na = F
)
}
\arguments{
\item{count_mat}{count matrix}

\item{thresholds}{vector of thresholds}

\item{folder}{folder to store the plots}

\item{cmap}{color map}

\item{include_na}{Whether to highlight entries not used for matrix factorization as missing}
}
\description{
Visualize binary masks and save as png files
}
