% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFBM_utils.R
\name{plot_error}
\alias{plot_error}
\title{Plot the mean absolute error against the number of thresholds}
\usage{
plot_error(performance)
}
\arguments{
\item{performance}{MAE of the same DFBM model with different number of thresholds}
}
\value{
a ggplot with number of thresholds on x axis and MAE on y axis
}
\description{
Plot the mean absolute error against the number of thresholds
}
