% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFBM_utils.R
\name{plot_prob_mat}
\alias{plot_prob_mat}
\title{Visualize probability matrices and save as png files}
\usage{
plot_prob_mat(
  result,
  folder,
  cmap = colorRamp2(c(0, 1), c("white", "#2c7cc7"))
)
}
\arguments{
\item{result}{DFBM result object}

\item{folder}{folder to store the plots}

\item{cmap}{color map}
}
\description{
Visualize probability matrices and save as png files
}
