% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFBM_utils.R
\name{plot_rank}
\alias{plot_rank}
\title{Visualize the optimal rank of matrix factorization for binary masks}
\usage{
plot_rank(result)
}
\arguments{
\item{result}{DFBM result object}
}
\value{
a ggplot with threshold on x axis and Rank on y axis
}
\description{
Visualize the optimal rank of matrix factorization for binary masks
}
\details{
For each binary masks from each threshold, the optimal rank for matrix factorization is different. Here I plot
the thresholds on the x axis and optimal ranks on the y axis
}
