% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RF.R
\name{repeat_rf}
\alias{repeat_rf}
\title{Randomly split the data into train/test and apply random forest model for prediction}
\usage{
repeat_rf(X, y, times = 20)
}
\arguments{
\item{X}{predictor matrix}

\item{y}{response variable}

\item{times}{how many train/test split to try}
}
\value{
a list with the training AUROC, test AUROC and the variable importance
}
\description{
Randomly split the data into train/test and apply random forest model for prediction
}
