% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_coda_lasso.R
\name{repeated_codalasso}
\alias{repeated_codalasso}
\title{Fit coda lasso multiple times with random train/test splitting. Cross validation involved}
\usage{
repeated_codalasso(y, X, times = 20)
}
\arguments{
\item{y}{response variable}

\item{X}{predictor variable matrix}

\item{times}{number of repeats of codalasso fitting}
}
\value{
list of training AUC, test AUC and coefficients of all runs
}
\description{
Fit coda lasso multiple times with random train/test splitting. Cross validation involved
}
