% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{save_heatmap}
\alias{save_heatmap}
\title{Plot heatmap}
\usage{
save_heatmap(
  X,
  entry_name,
  filename,
  rowannot = NULL,
  colannot = NULL,
  cmap,
  na_col = "#555555",
  border_col = "black",
  width = 10,
  height = 12,
  heatmap_ratio = 0.95,
  pad = 0.1,
  legend = F
)
}
\arguments{
\item{X}{input matrix}

\item{entry_name}{title of the colorbar}

\item{filename}{file name to store the heatmap}

\item{rowannot}{optional row annotation}

\item{colannot}{optional column annotation}

\item{cmap}{a ColorRamp object specifying the color bar}

\item{na_col}{color representing NA entries}

\item{border_col}{border color of the heatmap}

\item{width}{width of the heatmap, default 10cm}

\item{height}{height of the heatmap, default 12cm}

\item{heatmap_ratio}{proportion of canvas devoted to the heatmap, by default 0.95}

\item{pad}{padding when saving the figure, by default 0.1 cm}

\item{legend}{whether to show the colorbar or not}
}
\description{
Plot heatmap
}
