% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFBM_utils.R
\name{tune_dfbm}
\alias{tune_dfbm}
\title{Tune the performance of DFBM with different number of thresholds}
\usage{
tune_dfbm(
  X,
  rates = c(0.6, 0.7, 0.8, 0.85, 0.9, 0.93, 0.95),
  gap = 0.002,
  min_prop = 0.002,
  truth
)
}
\arguments{
\item{X}{count matrix}

\item{rates}{a vector of geometric sequence rates to tune for adjusting the number of thresholds. The higher the rate the more thresholds there are}

\item{gap}{minimum of the proportion of entries that change from one to zero between adjacent thresholds}

\item{min_prop}{minimum of the proportion of entries being larger than the largest threshold}

\item{truth}{true expected abundance matrix}
}
\value{
the 25\%, 50\% and 75\% percentile of median absolute error to the mean of each entry
}
\description{
Tune the performance of DFBM with different number of thresholds
}
