% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrywise_simulation.R
\name{ZIG_median}
\alias{ZIG_median}
\title{Median of zero inflated Gamma distribution}
\usage{
ZIG_median(zero_prob, mean_mat, scale = 1)
}
\arguments{
\item{zero_prob}{zero inflation probability}

\item{mean_mat}{mean matrix}

\item{scale}{gamma distribution scale}
}
\description{
Median of zero inflated Gamma distribution
}
