% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrywise_simulation.R
\name{ZIG_variance}
\alias{ZIG_variance}
\title{Variance of zero inflated gamma distribution}
\usage{
ZIG_variance(zero_prob, mean_mat, scale = 1)
}
\arguments{
\item{zero_prob}{zero inflation probability}

\item{mean_mat}{mean matrix}

\item{scale}{gamma distribution scale}
}
\description{
Variance of zero inflated gamma distribution
}
