% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_lasso.R
\name{fit_logistic_lasso}
\alias{fit_logistic_lasso}
\title{Fit lasso logistic regression}
\usage{
fit_logistic_lasso(X, y, seed = 1)
}
\arguments{
\item{X}{predictor matrix}

\item{y}{binary response vector}

\item{seed}{train test seed}
}
\description{
Fit lasso logistic regression
}
